requirements.txt
customtkinter
pillow < 10.0.0
ttkbootstrap
